requirements.txt
Django~=3.1.4
djangorestframework~=3.12.4
fastapi~=0.67.0
opencv-python~=4.5.2.52
pip~=21.1.1
Brotli~=1.0.9
cryptography~=3.3.2
numpy~=1.20.3
setuptools~=56.2.0
pytz~=2020.4
cffi~=1.14.4
certifi~=2020.12.5
chardet~=3.0.4
urllib3~=1.26.2
six~=1.15.0
idna~=2.10
sklearn~=0.0
scikit-learn~=0.24.2
requests~=2.25.0
hpack~=4.0.0
hyperframe~=6.0.1
h2~=4.0.0
h11~=0.12.0
Jinja2~=2.11.2
Werkzeug~=1.0.1
itsdangerous~=1.1.0
click~=7.1.2
python-dotenv~=0.19.0
blinker~=1.4
pyasn1~=0.4.8
urwid~=2.1.2
tornado~=6.1
MarkupSafe~=1.1.1
mitmproxy~=6.0.2
pyparsing~=2.4.7
wsproto~=1.0.0
asgiref~=3.4.1
passlib~=1.7.4
zstandard~=0.14.1
kaitaistruct~=0.9
publicsuffix2~=2.20191221
pyperclip~=1.8.2
sortedcontainers~=2.3.0
ldap3~=2.8.1
Flask~=1.1.4
msgpack~=1.0.2
pycparser~=2.20
Pygments~=2.7.3